moved the faqCard into the component folder
